refactor(table): move TableHeaderCell and TableHeaderRow to deprecated folder
Move deprecated TableHeaderCell and TableHeaderRow components
to the deprecated folder to better organize the codebase.
