+++ packages/react/src/table/deprecated/TableHeaderRow.js
@@ -1,9 +1,9 @@
 import React, { forwardRef } from 'react';
-import { Box } from '../box';
-import { useDefaultProps } from '../default-props';
-import { GROUP_VARIANT_HEADER, LAYOUT_TABLE } from './constants';
-import { useTableRowStyle } from './styles';
-import useTable from './useTable';
+import { Box } from '../../box';
+import { useDefaultProps } from '../../default-props';
+import { GROUP_VARIANT_HEADER, LAYOUT_TABLE } from '../constants';
+import { useTableRowStyle } from '../styles';
+import useTable from '../useTable';
 
 const TableHeaderRow = forwardRef((inProps, ref) => {
   const {
