+++ packages/react/src/table/index.js
@@ -1,13 +1,14 @@
 import { ResizeHandle as TableColumnResizeHandle } from '../resize-handle';
 import Table from './Table';
 import TableHeader from './TableHeader';
-import TableHeaderRow from './TableHeaderRow'; // deprecated
-import TableHeaderCell from './TableHeaderCell'; // deprecated
 import TableBody from './TableBody';
 import TableFooter from './TableFooter';
 import TableRow from './TableRow';
 import TableCell from './TableCell';
 import TableScrollbar from './TableScrollbar';
+
+import TableHeaderRow from './deprecated/TableHeaderRow'; // deprecated
+import TableHeaderCell from './deprecated/TableHeaderCell'; // deprecated
 
 Table.Header = TableHeader;
 Table.HeaderRow = TableHeaderRow; // deprecated
